optimizations:
1. wxLogXXX functions don't do anything if logging is suspended anyhow
2. wxColourDisplay() caches its result
3. wxRegConf is much, much more efficient
4. wxTextFile::Read() preallocs some memory
5. cache wxSystemSettings::GetSystemFont(wxSYS_DEFAULT_GUI_FONT);

Bug fix:

wxFont::GetTextExtent() wasn't using the given font


git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_2_BRANCH@9819 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/regconf.cpp
@@ -112,14 +112,28 @@ wxRegConfig::wxRegConfig(const wxString& appName, const wxString& vendorName,
   //else: we don't need to do all the complicated stuff above
 
   wxString str = strLocal.IsEmpty() ? strRoot : strLocal;
+
+  // as we're going to change the name of these keys fairly often and as
+  // there are only few of wxRegConfig objects (usually 1), we can allow
+  // ourselves to be generous and spend some memory to significantly improve
+  // performance of SetPath()
+  static const size_t MEMORY_PREALLOC = 512;
+
+  m_keyLocalRoot.ReserveMemoryForName(MEMORY_PREALLOC);
+  m_keyLocal.ReserveMemoryForName(MEMORY_PREALLOC);
+
   m_keyLocalRoot.SetName(wxRegKey::HKCU, SOFTWARE_KEY + str);
-  m_keyLocal.SetName(m_keyLocalRoot, "");
+  m_keyLocal.SetName(m_keyLocalRoot, _T(""));
 
   if ( bDoUseGlobal )
   {
     str = strGlobal.IsEmpty() ? strRoot : strGlobal;
+
+    m_keyGlobalRoot.ReserveMemoryForName(MEMORY_PREALLOC);
+    m_keyGlobal.ReserveMemoryForName(MEMORY_PREALLOC);
+
     m_keyGlobalRoot.SetName(wxRegKey::HKLM, SOFTWARE_KEY + str);
-    m_keyGlobal.SetName(m_keyGlobalRoot, "");
+    m_keyGlobal.SetName(m_keyGlobalRoot, _T(""));
   }
 
   // Create() will Open() if key already exists
@@ -145,50 +159,222 @@ wxRegConfig::~wxRegConfig()
 // ----------------------------------------------------------------------------
 // path management
 // ----------------------------------------------------------------------------
+
+// this function is called a *lot* of times (as I learned after seeing from
+// profiler output that it is called ~12000 times from Mahogany start up code!)
+// so it is important to optimize it - in particular, avoid using generic
+// string functions here and do everything manually because it is faster
+//
+// I still kept the old version to be able to check that the optimized code has
+// the same output as the non optimized version.
 void wxRegConfig::SetPath(const wxString& strPath)
 {
-  wxArrayString aParts;
+    // remember the old path
+    wxString strOldPath = m_strPath;
 
-  // because GetPath() returns "" when we're at root, we must understand
-  // empty string as "/"
-  if ( strPath.IsEmpty() || (strPath[0] == wxCONFIG_PATH_SEPARATOR) ) {
-    // absolute path
-    wxSplitPath(aParts, strPath);
-  }
-  else {
-    // relative path, combine with current one
-    wxString strFullPath = GetPath();
-    strFullPath << wxCONFIG_PATH_SEPARATOR << strPath;
-    wxSplitPath(aParts, strFullPath);
-  }
+#ifdef WX_DEBUG_SET_PATH // non optimized version kept here for testing
+    wxString m_strPathAlt;
 
-  // recombine path parts in one variable
-  wxString strOldPath = m_strPath, strRegPath;
-  m_strPath.Empty();
-  for ( size_t n = 0; n < aParts.Count(); n++ ) {
-    strRegPath << '\\' << aParts[n];
-    m_strPath << wxCONFIG_PATH_SEPARATOR << aParts[n];
-  }
+    {
+        wxArrayString aParts;
 
-  if ( m_strPath == strOldPath )
-      return;
+        // because GetPath() returns "" when we're at root, we must understand
+        // empty string as "/"
+        if ( strPath.IsEmpty() || (strPath[0] == wxCONFIG_PATH_SEPARATOR) ) {
+            // absolute path
+            wxSplitPath(aParts, strPath);
+        }
+        else {
+            // relative path, combine with current one
+            wxString strFullPath = GetPath();
+            strFullPath << wxCONFIG_PATH_SEPARATOR << strPath;
+            wxSplitPath(aParts, strFullPath);
+        }
 
-  // as we create the registry key when SetPath(key) is done, we can be left
-  // with plenty of empty keys if this was only done to try to read some value
-  // which, in fact, doesn't exist - to prevent this from happening we
-  // automatically delete the old key if it was empty
-  if ( m_keyLocal.IsEmpty() )
-  {
-      m_keyLocal.DeleteSelf();
-  }
+        // recombine path parts in one variable
+        wxString strRegPath;
+        m_strPathAlt.Empty();
+        for ( size_t n = 0; n < aParts.Count(); n++ ) {
+            strRegPath << '\\' << aParts[n];
+            m_strPathAlt << wxCONFIG_PATH_SEPARATOR << aParts[n];
+        }
+    }
+#endif // 0
 
-  // change current key(s)
-  m_keyLocal.SetName(m_keyLocalRoot, strRegPath);
-  m_keyGlobal.SetName(m_keyGlobalRoot, strRegPath);
-  m_keyLocal.Create();
+    // check for the most common case first
+    if ( strPath.empty() )
+    {
+        m_strPath = wxCONFIG_PATH_SEPARATOR;
+    }
+    else // not root
+    {
+        // construct the full path
+        wxString strFullPath;
+        if ( strPath[0u] == wxCONFIG_PATH_SEPARATOR )
+        {
+            // absolute path
+            strFullPath = strPath;
+        }
+        else // relative path
+        {
+            strFullPath.reserve(2*m_strPath.length());
 
-  wxLogNull nolog;
-  m_keyGlobal.Open();
+            strFullPath << m_strPath << wxCONFIG_PATH_SEPARATOR << strPath;
+        }
+
+        // simplify it: we need to handle ".." here
+
+        // count the total number of slashes we have to know if we can go upper
+        size_t totalSlashes = 0;
+
+        // position of the last slash to be able to backtrack to it quickly if
+        // needed, but we set it to -1 if we don't have a valid position
+        //
+        // we only remember the last position which means that we handle ".."
+        // quite efficiently but not "../.." - however the latter should be
+        // much more rare, so it is probably ok
+        int posLastSlash = -1;
+
+        const wxChar *src = strFullPath.c_str();
+        size_t len = strFullPath.length();
+        const wxChar *end = src + len;
+
+        wxChar *dst = m_strPath.GetWriteBuf(len);
+        wxChar *start = dst;
+
+        for ( ; src < end; src++, dst++ )
+        {
+            if ( *src == wxCONFIG_PATH_SEPARATOR )
+            {
+                // check for "/.."
+
+                // note that we don't have to check for src < end here as
+                // *end == 0 so can't be '.'
+                if ( src[1] == _T('.') && src[2] == _T('.') &&
+                     (src + 3 == end || src[3] == wxCONFIG_PATH_SEPARATOR) )
+                {
+                    if ( !totalSlashes )
+                    {
+                        wxLogWarning(_("'%s' has extra '..', ignored."),
+                                     strFullPath.c_str());
+                    }
+                    else // return to the previous path component
+                    {
+                        // do we already have its position?
+                        if ( posLastSlash == -1 )
+                        {
+                            // no, find it: note that we are sure to have one
+                            // because totalSlashes > 0 so we don't have to
+                            // check the boundary condition below
+
+                            // this is more efficient than strrchr()
+                            while ( *dst != wxCONFIG_PATH_SEPARATOR )
+                            {
+                                dst--;
+                            }
+                        }
+                        else // the position of last slash was stored
+                        {
+                            // go directly there
+                            dst = start + posLastSlash;
+
+                            // invalidate posLastSlash
+                            posLastSlash = -1;
+                        }
+
+                        // this shouldn't happen
+                        wxASSERT_MSG( *dst == wxCONFIG_PATH_SEPARATOR,
+                                      _T("error in wxRegConfig::SetPath") );
+
+                        // we killed one
+                        totalSlashes--;
+                    }
+
+                    // skip both dots
+                    src += 2;
+                }
+                else // not "/.."
+                {
+                    if ( (dst == start) || (dst[-1] != wxCONFIG_PATH_SEPARATOR) )
+                    {
+                        *dst = wxCONFIG_PATH_SEPARATOR;
+
+                        posLastSlash = dst - start;
+
+                        totalSlashes++;
+                    }
+                    //else: nothing to do, we squeeze several subseuquent
+                    //      slashes into one
+                }
+            }
+            else // normal character
+            {
+                // just copy
+                *dst = *src;
+            }
+        }
+
+        // NUL terminate the string
+        if ( dst[-1] == wxCONFIG_PATH_SEPARATOR && (dst != start + 1) )
+        {
+            // if it has a trailing slash we remove it unless it is the only
+            // string character
+            dst--;
+        }
+
+        *dst = _T('\0');
+
+        m_strPath.UngetWriteBuf(dst - start);
+    }
+
+#ifdef WX_DEBUG_SET_PATH
+    wxASSERT( m_strPath == m_strPathAlt );
+#endif
+
+    if ( m_strPath == strOldPath )
+        return;
+
+    // registry APIs want backslashes instead of slashes
+    wxString strRegPath;
+    size_t len = m_strPath.length();
+
+    const wxChar *src = m_strPath.c_str();
+    wxChar *dst = strRegPath.GetWriteBuf(len);
+
+    const wxChar *end = src + len;
+    for ( ; src < end; src++, dst++ )
+    {
+        if ( *src == wxCONFIG_PATH_SEPARATOR )
+            *dst = _T('\\');
+        else
+            *dst = *src;
+    }
+
+    strRegPath.UngetWriteBuf(len);
+
+    // this is not needed any longer as we don't create keys unnecessarily any
+    // more (now it is done on demand, i.e. only when they're going to contain
+    // something)
+#if 0
+    // as we create the registry key when SetPath(key) is done, we can be left
+    // with plenty of empty keys if this was only done to try to read some
+    // value which, in fact, doesn't exist - to prevent this from happening we
+    // automatically delete the old key if it was empty
+    if ( m_keyLocal.Exists() && LocalKey().IsEmpty() )
+    {
+        m_keyLocal.DeleteSelf();
+    }
+#endif // 0
+
+    // change current key(s)
+    m_keyLocal.SetName(m_keyLocalRoot, strRegPath);
+    m_keyGlobal.SetName(m_keyGlobalRoot, strRegPath);
+
+    // don't create it right now, wait until it is accessed
+    //m_keyLocal.Create();
+
+    wxLogNull nolog;
+    m_keyGlobal.Open();
 }
 
 // ----------------------------------------------------------------------------
@@ -218,7 +404,7 @@ bool wxRegConfig::GetNextGroup(wxString& str, long& lIndex) const
   if ( m_keyGlobal.IsOpened() && !IS_LOCAL_INDEX(lIndex) ) {
     // try to find a global entry which doesn't appear locally
     while ( m_keyGlobal.GetNextKey(str, lIndex) ) {
-      if ( !m_keyLocal.HasSubKey(str) ) {
+      if ( !m_keyLocal.Exists() || !LocalKey().HasSubKey(str) ) {
         // ok, found one - return it
         return TRUE;
       }
@@ -228,10 +414,14 @@ bool wxRegConfig::GetNextGroup(wxString& str, long& lIndex) const
     lIndex |= LOCAL_MASK;
   }
 
+  // if we don't have the key at all, don't try to enumerate anything under it
+  if ( !m_keyLocal.Exists() )
+      return FALSE;
+
   // much easier with local entries: get the next one we find
   // (don't forget to clear our flag bit and set it again later)
   lIndex &= ~LOCAL_MASK;
-  bool bOk = m_keyLocal.GetNextKey(str, lIndex);
+  bool bOk = LocalKey().GetNextKey(str, lIndex);
   lIndex |= LOCAL_MASK;
 
   return bOk;
@@ -249,7 +439,7 @@ bool wxRegConfig::GetNextEntry(wxString& str, long& lIndex) const
   if ( m_keyGlobal.IsOpened() && !IS_LOCAL_INDEX(lIndex) ) {
     // try to find a global entry which doesn't appear locally
     while ( m_keyGlobal.GetNextValue(str, lIndex) ) {
-      if ( !m_keyLocal.HasValue(str) ) {
+      if ( !m_keyLocal.Exists() || !LocalKey().HasValue(str) ) {
         // ok, found one - return it
         return TRUE;
       }
@@ -259,10 +449,14 @@ bool wxRegConfig::GetNextEntry(wxString& str, long& lIndex) const
     lIndex |= LOCAL_MASK;
   }
 
+  // if we don't have the key at all, don't try to enumerate anything under it
+  if ( !m_keyLocal.Exists() )
+      return FALSE;
+
   // much easier with local entries: get the next one we find
   // (don't forget to clear our flag bit and set it again later)
   lIndex &= ~LOCAL_MASK;
-  bool bOk = m_keyLocal.GetNextValue(str, lIndex);
+  bool bOk = LocalKey().GetNextValue(str, lIndex);
   lIndex |= LOCAL_MASK;
 
   return bOk;
@@ -312,7 +506,8 @@ bool wxRegConfig::HasGroup(const wxString& key) const
 
     wxString strName(path.Name());
 
-    return m_keyLocal.HasSubKey(strName) || m_keyGlobal.HasSubKey(strName);
+    return (m_keyLocal.Exists() && LocalKey().HasSubKey(strName)) ||
+           m_keyGlobal.HasSubKey(strName);
 }
 
 bool wxRegConfig::HasEntry(const wxString& key) const
@@ -321,7 +516,8 @@ bool wxRegConfig::HasEntry(const wxString& key) const
 
     wxString strName(path.Name());
 
-    return m_keyLocal.HasValue(strName) || m_keyGlobal.HasValue(strName);
+    return (m_keyLocal.Exists() && LocalKey().HasValue(strName)) ||
+           m_keyGlobal.HasValue(strName);
 }
 
 wxConfigBase::EntryType wxRegConfig::GetEntryType(const wxString& key) const
@@ -331,7 +527,7 @@ wxConfigBase::EntryType wxRegConfig::GetEntryType(const wxString& key) const
     wxString strName(path.Name());
 
     bool isNumeric;
-    if ( m_keyLocal.HasValue(strName) )
+    if ( m_keyLocal.Exists() && LocalKey().HasValue(strName) )
         isNumeric = m_keyLocal.IsNumericValue(strName);
     else if ( m_keyGlobal.HasValue(strName) )
         isNumeric = m_keyGlobal.IsNumericValue(strName);
@@ -355,7 +551,7 @@ bool wxRegConfig::Read(const wxString& key, wxString *pStr) const
   // overriden by the local key with the same name
   if ( IsImmutable(path.Name()) ) {
     if ( TryGetValue(m_keyGlobal, path.Name(), *pStr) ) {
-      if ( m_keyLocal.HasValue(path.Name()) ) {
+      if ( m_keyLocal.Exists() && LocalKey().HasValue(path.Name()) ) {
         wxLogWarning(wxT("User value for immutable key '%s' ignored."),
                    path.Name().c_str());
       }
@@ -369,7 +565,7 @@ bool wxRegConfig::Read(const wxString& key, wxString *pStr) const
   }
 
   // first try local key
-  if ( TryGetValue(m_keyLocal, path.Name(), *pStr) ||
+  if ( (m_keyLocal.Exists() && TryGetValue(LocalKey(), path.Name(), *pStr)) ||
        (bQueryGlobal && TryGetValue(m_keyGlobal, path.Name(), *pStr)) ) {
     // nothing to do
 
@@ -391,7 +587,7 @@ bool wxRegConfig::Read(const wxString& key, wxString *pStr,
   // overriden by the local key with the same name
   if ( IsImmutable(path.Name()) ) {
     if ( TryGetValue(m_keyGlobal, path.Name(), *pStr) ) {
-      if ( m_keyLocal.HasValue(path.Name()) ) {
+      if ( m_keyLocal.Exists() && LocalKey().HasValue(path.Name()) ) {
         wxLogWarning(wxT("User value for immutable key '%s' ignored."),
                    path.Name().c_str());
       }
@@ -405,7 +601,7 @@ bool wxRegConfig::Read(const wxString& key, wxString *pStr,
   }
 
   // first try local key
-  if ( TryGetValue(m_keyLocal, path.Name(), *pStr) ||
+  if ( (m_keyLocal.Exists() && TryGetValue(LocalKey(), path.Name(), *pStr)) ||
        (bQueryGlobal && TryGetValue(m_keyGlobal, path.Name(), *pStr)) ) {
     *pStr = wxConfigBase::ExpandEnvVars(*pStr);
     return TRUE;
@@ -434,7 +630,7 @@ bool wxRegConfig::Read(const wxString& key, long *plResult) const
   // overriden by the local key with the same name
   if ( IsImmutable(path.Name()) ) {
     if ( TryGetValue(m_keyGlobal, path.Name(), plResult) ) {
-      if ( m_keyLocal.HasValue(path.Name()) ) {
+      if ( m_keyLocal.Exists() && LocalKey().HasValue(path.Name()) ) {
         wxLogWarning(wxT("User value for immutable key '%s' ignored."),
                      path.Name().c_str());
       }
@@ -448,7 +644,7 @@ bool wxRegConfig::Read(const wxString& key, long *plResult) const
   }
 
   // first try local key
-  if ( TryGetValue(m_keyLocal, path.Name(), plResult) ||
+  if ( (m_keyLocal.Exists() && TryGetValue(LocalKey(), path.Name(), plResult)) ||
        (bQueryGlobal && TryGetValue(m_keyGlobal, path.Name(), plResult)) ) {
     return TRUE;
   }
@@ -464,7 +660,7 @@ bool wxRegConfig::Write(const wxString& key, const wxString& szValue)
     return FALSE;
   }
 
-  return m_keyLocal.SetValue(path.Name(), szValue);
+  return LocalKey().SetValue(path.Name(), szValue);
 }
 
 bool wxRegConfig::Write(const wxString& key, long lValue)
@@ -476,7 +672,7 @@ bool wxRegConfig::Write(const wxString& key, long lValue)
     return FALSE;
   }
 
-  return m_keyLocal.SetValue(path.Name(), lValue);
+  return LocalKey().SetValue(path.Name(), lValue);
 }
 
 // ----------------------------------------------------------------------------
@@ -547,13 +743,15 @@ bool wxRegConfig::DeleteEntry(const wxString& value, bool bGroupIfEmptyAlso)
 {
   wxConfigPathChanger path(this, value);
 
-  if ( !m_keyLocal.DeleteValue(path.Name()) )
-    return FALSE;
+  if ( m_keyLocal.Exists() ) {
+    if ( !m_keyLocal.DeleteValue(path.Name()) )
+      return FALSE;
 
-  if ( m_keyLocal.IsEmpty() ) {
-    wxString strKey = GetPath().AfterLast(wxCONFIG_PATH_SEPARATOR);
-    SetPath("..");  // changes m_keyLocal
-    return m_keyLocal.DeleteKey(strKey);
+    if ( m_keyLocal.IsEmpty() ) {
+      wxString strKey = GetPath().AfterLast(wxCONFIG_PATH_SEPARATOR);
+      SetPath("..");  // changes m_keyLocal
+      return LocalKey().DeleteKey(strKey);
+    }
   }
 
   return TRUE;
@@ -563,7 +761,7 @@ bool wxRegConfig::DeleteGroup(const wxString& key)
 {
   wxConfigPathChanger path(this, key);
 
-  return m_keyLocal.DeleteKey(path.Name());
+  return m_keyLocal.Exists() ? LocalKey().DeleteKey(path.Name()) : TRUE;
 }
 
 bool wxRegConfig::DeleteAll()
